# docs/swagger.yaml
basePath: /v1
definitions:
...
  api.createPayOutRequest:
    properties:
      amount:
        description: |-
          Amount of the pay out.
          Required: true
          example: 1000
          in: body
        type: number
      owner:
        description: |-
          Owner of the pay out.
          Required: true
          example: john_doe
          in: body
        type: string
      subject:
        description: |-
          Subject of the pay out.
          Required: true
          example: pay_out1
          in: body
        type: string
    required:
    - amount
    - owner
    - subject
    type: object
...
  api.errorResponse:
    properties:
      error:
        type: string
    type: object
  api.listRequest:
    properties:
      page_id:
        description: |-
          PageID is the page number.
          Required: true
          example: 1
          in: body
          minimum: 1
        minimum: 1
        type: integer
      page_size:
        description: |-
          PageSize is the number of projects per page.
          Required: true
          example: 5
          in: body
          minimum: 5
          maximum: 10
        maximum: 100
        minimum: 5
        type: integer
    required:
    - page_id
    - page_size
    type: object
...
  api.searchRequest:
    properties:
      page_id:
        description: |-
          PageID is the page number.
          Required: true
          example: 1
          in: body
          minimum: 1
        minimum: 1
        type: integer
      page_size:
        description: |-
          PageSize is the number of projects per page.
          Required: true
          example: 5
          in: body
          minimum: 5
          maximum: 10
        maximum: 10
        minimum: 5
        type: integer
      query:
        description: |-
          Query is the search query.
          Required: true
          example: project1
          in: body
        type: string
    required:
    - page_id
    - page_size
    - query
    type: object
...
  db.PayOut:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      owner:
        type: string
      subject:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: lushenle@gmail.com
    name: API Support
    url: http://localhost:8080/api/v1/support
  description: This is a sample server for PLAM API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://localhost:8080/api/v1/terms/
  title: PLAM API
  version: "1.0"
paths:
...
  /pay_outs:
    post:
      consumes:
      - application/json
      description: Create a new pay out.
      parameters:
      - description: Create Pay Out Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createPayOutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Pay Out created
          schema:
            $ref: '#/definitions/db.PayOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a pay out
      tags:
      - pay_outs
  /pay_outs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a pay out by ID.
      parameters:
      - description: Pay Out ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Pay Out deleted
          schema:
            $ref: '#/definitions/db.PayOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a pay out
      tags:
      - pay_outs
    get:
      consumes:
      - application/json
      description: Get a pay out by ID.
      parameters:
      - description: Pay Out ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Pay Out found
          schema:
            $ref: '#/definitions/db.PayOut'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a pay out
      tags:
      - pay_outs
  /pay_outs/all:
    post:
      consumes:
      - application/json
      description: List all pay outs.
      parameters:
      - description: List Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.listRequest'
      produces:
      - application/json
      responses:
        "200":
          description: List of pay outs
          schema:
            items:
              $ref: '#/definitions/db.PayOut'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all pay outs
      tags:
      - pay_outs
  /pay_outs/search:
    post:
      consumes:
      - application/json
      description: Search pay outs by owner.
      parameters:
      - description: Search Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.searchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Pay Outs found
          schema:
            items:
              $ref: '#/definitions/db.PayOut'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search pay outs
      tags:
      - pay_outs
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  apiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"